feat: Implementation of Han et al.'s 2020 Hyperparameters
This commit introduces changes to the codebase to incorporate the
hyperparameters outlined in the Han et al. 2020 article. In particular,
adjustments have been made in the `variant.py` file to ensure alignment
with the parameters discussed in the article. Furthermore, for the sake
of numerical stability, the initial value of $\lambda$ has been set to
0.99.